fix more linting errors
