+++ client/.eslintrc.json
@@ -18,7 +18,7 @@
         }
     },
     "rules": {
-        "no-underscore-dangle": [2, { "allowAfterThis": true }],
+        "no-underscore-dangle": "off",
         "no-param-reassign": ["error", { "props": false }],
         "react/function-component-definition": [
             2,
@@ -28,6 +28,9 @@
             }
         ],
         "no-console": "off",
-        "@typescript-eslint/no-unused-vars" : "off"
+        "@typescript-eslint/no-unused-vars" : "off",
+        "no-nested-ternary": "off",
+        "import/prefer-default-export": "off",
+        "react/require-default-props": "off"
     }
 }
