+++ client/package.json
@@ -8,6 +8,7 @@
         "@mui/icons-material": "^5.1.1",
         "@mui/material": "^5.1.1",
         "@mui/styles": "^5.1.1",
+        "@mui/system": "^5.4.0",
         "@svgdotjs/svg.js": "^3.1.1",
         "@testing-library/jest-dom": "^5.15.1",
         "@testing-library/react": "^11.1.0",
@@ -23,11 +24,12 @@
         "framer-motion": "^4.1.17",
         "identity-obj-proxy": "^3.0.0",
         "marked": "^2.1.3",
+        "prop-types": "^15.8.1",
         "react": "^17.0.2",
-        "react-cookie": "^4.1.1",
         "react-code-blocks": "0.0.9-0",
         "react-collapse-pane": "^2.0.1",
         "react-console-emulator": "^5.0.1",
+        "react-cookie": "^4.1.1",
         "react-dom": "^17.0.2",
         "react-helmet": "^6.1.0",
         "react-intersection-observer": "^8.32.5",
