+++ client/src/components/Tabs/Tabs.tsx
@@ -2,7 +2,7 @@ import {
   Box, Tab, Tabs, Theme, Typography,
 } from '@mui/material';
 import { useTheme } from '@mui/styles';
-import { FC, useState } from 'react';
+import React, { FC, useState } from 'react';
 import { Topic } from 'utils/apiRequests';
 import TabRenderer from './TabRenderer';
 import './Tabs.scss';
