+++ client/src/components/Visualiser/Controller/GUIMode/GUIMode.tsx
@@ -1,5 +1,5 @@
 import { Box } from '@mui/material';
-import { FC } from 'react';
+import React, { FC } from 'react';
 import OperationsTree from './GUIOperations/OperationsTree';
 
 interface Props {
