+++ client/src/components/Visualiser/Controller/Terminal/Terminal.tsx
@@ -107,9 +107,9 @@ const Terminal: FC<Props> = ({ executeCommand, topicTitle }) => {
             cursor={{ hideWhenDone: true }}
             avgTypingDelay={30}
           >
-            Type "man" to see commands available for '
+            Type &quotman&quot to see commands available for &apos
             {topicTitle}
-            '.
+            &apos.
           </Typist>
         )}
       </div>
