+++ client/src/components/Visualiser/Controller/index.ts
@@ -1,1 +1,3 @@
+/* eslint-disable import/prefer-default-export */
+
 export { default as VisualiserController } from './VisualiserController';
