+++ client/src/components/Visualiser/index.ts
@@ -1,3 +1,1 @@
-/* eslint-disable import/prefer-default-export */
-
 export { default as Visualiser } from './Visualiser';
